added distributed general population mortality method and some plots
This adds a method to capture gpop OS more accurately using a parametric
distribution of baseline age (normal or lognormal) and extrapolating
gpop mortality for each age from 0 to 99 at baseline to the end of the
model time horizon, with optional death on 100th birthday. It then takes
these lines and weights them by baselien proportion at each age and sums
columns to give an OS line. This can then be used in the same way as the
"typical" method but captures the non-linearity in mortality given age
in the life table.

It then uses the DSU lookup table for utility to produce absolute and
relative utiltiy per sex taking the distributed OS into account for
every cycle of a model, thereby correcting for gpop utiltiy as well (for
the multiplicative and minima approaches)

Finally, some interesting plots are produced